Review fixes
- add delay to tests
- add comments
- rename OpTask to ExecNodeTask
- separate OpTask and OutputTask as subclasses of ExecNodeTask
